Develop to beta

+++ podchat/src/main/java/com/fanap/podchat/persistance/MessageDatabaseHelper.java
@@ -50,6 +50,7 @@ import com.fanap.podchat.model.ChatResponse;
 import com.fanap.podchat.model.ConversationSummery;
 import com.fanap.podchat.model.ReplyInfoVO;
 import com.fanap.podchat.chat.pin.pin_message.model.ResultPinMessage;
+import com.fanap.podchat.model.ResultHistory;
 import com.fanap.podchat.persistance.dao.MessageDao;
 import com.fanap.podchat.persistance.dao.MessageQueueDao;
 import com.fanap.podchat.requestobject.RequestGetHistory;
@@ -73,6 +74,7 @@ import java.util.Locale;
 
 import javax.inject.Inject;
 
+
 @SuppressWarnings("unchecked")
 public class MessageDatabaseHelper {
 
@@ -95,25 +97,38 @@ public class MessageDatabaseHelper {
     public void clearAllData(Chat.IClearMessageCache listener) {
 
         try {
-            new java.lang.Thread(() -> {
 
-                try {
 
-                    appDatabase.clearAllTables();
+            Runnable clearTablesTask = () -> appDatabase.clearAllTables();
 
-                    java.lang.Thread.sleep(2000);
+            Runnable vacuumTask = () -> messageDao.vacuumDb(new SimpleSQLiteQuery("VACUUM"));
 
-                    messageDao.vacuumDb(new SimpleSQLiteQuery("VACUUM"));
+            new PodThreadManager()
+                    .addNewTask(clearTablesTask)
+                    .addNewTask(vacuumTask)
+                    .addNewTask(listener::onCacheDatabaseCleared)
+                    .runTasksSynced();
 
-                    java.lang.Thread.sleep(1000);
 
-                    listener.onCacheDatabaseCleared();
-
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                    listener.onExceptionOccurred(e.getMessage());
-                }
-            }).start();
+//            new java.lang.Thread(() -> {
+//
+//                try {
+//
+//                    appDatabase.clearAllTables();
+//
+//                    java.lang.Thread.sleep(2000);
+//
+//                    messageDao.vacuumDb(new SimpleSQLiteQuery("VACUUM"));
+//
+//                    java.lang.Thread.sleep(1000);
+//
+//                    listener.onCacheDatabaseCleared();
+//
+//                } catch (InterruptedException e) {
+//                    e.printStackTrace();
+//                    listener.onExceptionOccurred(e.getMessage());
+//                }
+//            }).start();
         } catch (Exception e) {
             e.printStackTrace();
             listener.onExceptionOccurred(e.getMessage());
@@ -130,7 +145,7 @@ public class MessageDatabaseHelper {
         SQLCipherUtils.decrypt(context, file, passphrase);
     }
 
-    public void reEcryptKey(char[] passphrase) {
+    public void reEncryptKey(char[] passphrase) {
         SupportSQLiteOpenHelper supportSQLiteOH = appDatabase.getOpenHelper();
         supportSQLiteOH.getWritableDatabase();
         SupportSQLiteDatabase supportSQLiteDatabase = supportSQLiteOH.getReadableDatabase();
@@ -410,9 +425,31 @@ public class MessageDatabaseHelper {
         worker(() -> messageQueueDao.deleteWaitMessageQueue(uniqueId));
     }
 
-    //TODO get from another thread
     public List<WaitQueueCache> getAllWaitQueueMsg() {
         return messageQueueDao.getAllWaitQueueMsg();
+    }
+
+    public void getWaitQueueUniqueIdList(OnWorkDone listener) {
+
+        List<String> items = new ArrayList<>();
+
+        new PodThreadManager()
+                .doThisSafe(
+                        () -> items.addAll(messageQueueDao.getAllWaitQueueMsgUniqueId()),
+                        new PodThreadManager.IComplete() {
+                            @Override
+                            public void onComplete() {
+
+                                listener.onWorkDone(items);
+                            }
+
+                            @Override
+                            public void onError(String error) {
+
+                                listener.onWorkDone(null);
+                            }
+                        });
+
     }
 
     @Nullable
@@ -426,7 +463,7 @@ public class MessageDatabaseHelper {
         return null;
     }
 
-    //TODO get from another thread
+
     @NonNull
     public List<Failed> getAllWaitQueueCacheByThreadId(long threadId) {
         List<Failed> listQueues = new ArrayList<>();
@@ -469,12 +506,43 @@ public class MessageDatabaseHelper {
         worker(() -> messageQueueDao.deleteSendingMessageQueue(uniqueId));
     }
 
-    //TODO get from another thread
-    public SendingQueueCache getSendingQueueCache(String uniqueId) {
+
+    private SendingQueueCache getSendingQueueCache(String uniqueId) {
         return messageQueueDao.getSendingQueue(uniqueId);
     }
 
-    //TODO get from another thread
+    public void moveFromSendQueueToWaitQueue(String uniqueId) {
+
+        new PodThreadManager().doThisSafe(() -> {
+            SendingQueueCache sendingQueue = getSendingQueueCache(uniqueId);
+
+            deleteSendingMessageQueue(uniqueId);
+
+            insertWaitMessageQueue(sendingQueue);
+        }, new PodThreadManager.IComplete() {
+            @Override
+            public void onComplete() {
+
+            }
+
+            @Override
+            public void onError(String error) {
+
+            }
+        });
+
+    }
+
+    public void moveFromWaitQueueToSendQueue(String uniqueId, OnWorkDone listener) {
+
+        new PodThreadManager().doThisAndGo(() -> {
+            SendingQueueCache sendingQueue = getWaitQueueMsgByUnique(uniqueId);
+            deleteWaitQueueMsgs(uniqueId);
+            insertSendingMessageQueue(sendingQueue);
+            listener.onWorkDone(sendingQueue);
+        });
+
+    }
 
     public List<SendingQueueCache> getAllSendingQueue() {
         return messageQueueDao.getAllSendingQueue();
@@ -485,10 +553,8 @@ public class MessageDatabaseHelper {
     }
 
 
-    //TODO get from another thread
-
     @NonNull
-    public SendingQueueCache getWaitQueueMsgByUnique(String uniqueId) {
+    private SendingQueueCache getWaitQueueMsgByUnique(String uniqueId) {
         SendingQueueCache sendingQueueCache = new SendingQueueCache();
         WaitQueueCache waitQueueCache = messageQueueDao.getWaitQueueMsgByUniqueId(uniqueId);
 
@@ -513,12 +579,12 @@ public class MessageDatabaseHelper {
         return messageQueueDao.getAllSendingQueue();
     }
 
-    //TODO get from another thread
+
     @NonNull
     public List<Sending> getAllSendingQueueByThreadId(long threadId) {
 
         List<Sending> listQueues = new ArrayList<>();
-        List<SendingQueueCache> listCaches = messageQueueDao.getAllSendingQueue();
+        List<SendingQueueCache> listCaches = messageQueueDao.getAllSendingQueueByThredId(threadId);
 
         for (SendingQueueCache queueCache : listCaches) {
             Sending sending = new Sending();
@@ -548,7 +614,6 @@ public class MessageDatabaseHelper {
         worker(() -> messageQueueDao.insertUploadingQueue(uploadingQueue));
     }
 
-    //TODO get from another thread
     @NonNull
     public List<Uploading> getAllUploadingQueueByThreadId(long threadId) {
         List<Uploading> uploadingQueues = new ArrayList<>();
@@ -578,7 +643,6 @@ public class MessageDatabaseHelper {
         worker(() -> messageQueueDao.deleteUploadingQueue(uniqueId));
     }
 
-    //TODO get from another thread
     public UploadingQueueCache getUploadingQ(String uniqueId) {
         return messageQueueDao.getUploadingQ(uniqueId);
     }
@@ -596,45 +660,49 @@ public class MessageDatabaseHelper {
         //then set previous message as last message
 
         worker(() -> {
-            ThreadVo threadVo = messageDao.getThreadById(subjectId);
 
-            if (threadVo != null) {
+            if (subjectId > 0) {
 
-                long threadLastMessageId = threadVo.getLastMessageVOId();
+                ThreadVo threadVo = messageDao.getThreadById(subjectId);
 
-                if (threadLastMessageId == id) {
+                if (threadVo != null) {
 
-                    //this is last message
-                    List<CacheMessageVO> cacheMessage = messageDao.getMessage(id);
+                    long threadLastMessageId = threadVo.getLastMessageVOId();
 
-                    if (!Util.isNullOrEmpty(cacheMessage)) {
+                    if (threadLastMessageId == id) {
 
-                        long previousMessageId = cacheMessage.get(0).getPreviousId();
+                        //this is last message
+                        List<CacheMessageVO> cacheMessage = messageDao.getMessage(id);
 
-                        //Get previous message
-                        List<CacheMessageVO> previousMessage = messageDao.getMessage(previousMessageId);
+                        if (!Util.isNullOrEmpty(cacheMessage)) {
 
-                        if (!Util.isNullOrEmpty(previousMessage)) {
+                            long previousMessageId = cacheMessage.get(0).getPreviousId();
 
-                            String message = previousMessage.get(0).getMessage();
-                            messageDao.updateThreadLastMessageVOId(subjectId, previousMessageId, message);
+                            //Get previous message
+                            List<CacheMessageVO> previousMessage = messageDao.getMessage(previousMessageId);
 
-                        } else {
+                            if (!Util.isNullOrEmpty(previousMessage)) {
 
-                            // this thread has only one message
-                            messageDao.removeThreadLastMessageVO(subjectId);
+                                String message = previousMessage.get(0).getMessage();
+                                messageDao.updateThreadLastMessageVOId(subjectId, previousMessageId, message);
+
+                            } else {
+
+                                // this thread has only one message
+                                messageDao.removeThreadLastMessageVO(subjectId);
+                            }
                         }
                     }
                 }
-            }
 
+                //delete from pinned message
+                PinMessageVO pinnedMessage = messageDao.getThreadPinnedMessage(subjectId);
 
-            //delete from pinned message
-            PinMessageVO pinnedMessage = messageDao.getThreadPinnedMessage(subjectId);
+                if (pinnedMessage != null && pinnedMessage.getMessageId() == id) {
 
-            if (pinnedMessage != null && pinnedMessage.getMessageId() == id) {
+                    messageDao.deletePinnedMessageById(id);
 
-                messageDao.deletePinnedMessageById(id);
+                }
 
             }
 
@@ -969,7 +1037,6 @@ public class MessageDatabaseHelper {
             order = "desc";
         }
 
-        //TODO
         String rawQuery = "DELETE FROM CacheMessageVO WHERE threadVoId =" + threadId + " ORDER BY timeStamp desc LIMIT 50 OFFSET 0";
 
 //        rawQuery = rawQuery + " LIMIT " + count + " OFFSET " + offset;
@@ -987,7 +1054,6 @@ public class MessageDatabaseHelper {
     }
 
 
-    //todo get from another thread
     @NonNull
     public List<MessageVO> getHistories(@NonNull History history, long threadId) {
 
@@ -1030,6 +1096,85 @@ public class MessageDatabaseHelper {
 
 
         return messageVOS;
+    }
+
+    public void getHistories(@NonNull History history, long threadId, OnWorkDone listener) {
+
+
+        List<MessageVO> messageVOS = new ArrayList<>();
+        List<CacheMessageVO> cacheMessageVOS;
+        long fromTime = history.getFromTime();
+        long fromTimeNanos = history.getFromTimeNanos();
+        long toTime = history.getToTime();
+        long toTimeNanos = history.getToTimeNanos();
+        long messageId = history.getId();
+        long offset = history.getOffset();
+        long count = history.getCount();
+        String query = history.getQuery();
+        String order = history.getOrder();
+        offset = offset >= 0 ? offset : 0;
+        count = count > 0 ? count : 50;
+        if (Util.isNullOrEmpty(order)) {
+            order = "desc";
+        }
+
+        String rawQuery = "SELECT * FROM CacheMessageVO WHERE threadVoId =" + threadId;
+
+        rawQuery = addMessageIdIfExist(messageId, rawQuery);
+
+        rawQuery = addFromTimeIfExist(fromTime, fromTimeNanos, rawQuery);
+
+        rawQuery = addToTimeIfExist(toTime, toTimeNanos, rawQuery);
+
+        rawQuery = addQueryIfExist(query, rawQuery);
+
+        rawQuery = addOrderAndLimitAndOffset(offset, count, order, rawQuery);
+
+        SupportSQLiteQuery sqLiteQuery = new SimpleSQLiteQuery(rawQuery);
+
+        cacheMessageVOS = messageDao.getRawHistory(sqLiteQuery);
+
+
+        prepareMessageVOs(messageVOS, cacheMessageVOS);
+
+
+        long contentCount = getHistoryContentCount(threadId);
+
+        List<Sending> sendingList = getAllSendingQueueByThreadId(threadId);
+
+        List<Uploading> uploadingList = getAllUploadingQueueByThreadId(threadId);
+
+        List<Failed> failedList = getAllWaitQueueCacheByThreadId(threadId);
+
+        ChatResponse<ResultHistory> chatResponse = new ChatResponse<>();
+        chatResponse.setCache(true);
+
+        ResultHistory resultHistory = new ResultHistory();
+        resultHistory.setHistory(messageVOS);
+
+        resultHistory.setNextOffset(history.getOffset() + messageVOS.size());
+        resultHistory.setContentCount(contentCount);
+        if (messageVOS.size() + history.getOffset() < contentCount) {
+            resultHistory.setHasNext(true);
+        } else {
+            resultHistory.setHasNext(false);
+        }
+
+        resultHistory.setHistory(messageVOS);
+
+
+        resultHistory.setSending(sendingList);
+        resultHistory.setUploadingQueue(uploadingList);
+        resultHistory.setFailed(failedList);
+        chatResponse.setErrorCode(0);
+        chatResponse.setHasError(false);
+        chatResponse.setErrorMessage("");
+        chatResponse.setResult(resultHistory);
+        chatResponse.setCache(true);
+        chatResponse.setSubjectId(threadId);
+
+
+        listener.onWorkDone(chatResponse);
     }
 
 
@@ -1320,19 +1465,16 @@ public class MessageDatabaseHelper {
     }
 
 
-    //TODO generate two method and get from another thread in one
     public List<CacheMessageVO> getMessageById(long messageId) {
         return messageDao.getMessage(messageId);
     }
 
 
-    //todo could load in another thread
     public long getHistoryContentCount(long threadVoId) {
         return messageDao.getHistoryCount(threadVoId);
     }
 
 
-    //todo get from another thread ***
     @NonNull
     public List<Contact> getContacts(Integer count, Long offset) {
 
@@ -1380,7 +1522,6 @@ public class MessageDatabaseHelper {
         return contacts;
     }
 
-    //todo could load in another thread
     public int getContactCount() {
         return messageDao.getContactCount();
     }
@@ -1572,7 +1713,6 @@ public class MessageDatabaseHelper {
     }
 
 
-    //todo get from another thread
     @NonNull
     public List<BlockedContact> getBlockedContacts(Long count, Long offset) {
 
@@ -1715,7 +1855,6 @@ public class MessageDatabaseHelper {
 
     }
 
-    //todo get from another thread
     public UserInfo getUserInfo() {
 
         UserInfo userInfo = new UserInfo();
@@ -1733,7 +1872,6 @@ public class MessageDatabaseHelper {
         return userInfo;
     }
 
-    //todo could load from another thread
     public int getThreadCount() {
         return messageDao.getThreadCount();
     }
@@ -1789,6 +1927,9 @@ public class MessageDatabaseHelper {
             if (threadVos != null) {
 
                 for (ThreadVo threadVo : threadVos) {
+
+                    if (threadVo.getId() == 0)
+                        continue;
 
                     CacheParticipant cacheParticipant;
                     CacheReplyInfoVO cacheReplyInfoVO;
@@ -2058,7 +2199,7 @@ public class MessageDatabaseHelper {
     @NonNull
     public void getThreadIdsList(OnWorkDone listener) {
 
-        new java.lang.Thread(() -> {
+        worker(() -> {
 
             String rawQuery = "select id from ThreadVo";
 
@@ -2068,7 +2209,7 @@ public class MessageDatabaseHelper {
 
             listener.onWorkDone(tvo);
 
-        }).start();
+        });
 
 
     }
@@ -2183,7 +2324,8 @@ public class MessageDatabaseHelper {
 
     public void saveNewThread(Thread thread) {
 
-        worker(() -> prepareThreadVOAndSaveIt(thread));
+        if (thread != null && thread.getId() > 0)
+            worker(() -> prepareThreadVOAndSaveIt(thread));
 
     }
 
@@ -2194,7 +2336,8 @@ public class MessageDatabaseHelper {
         worker(() -> {
 
             for (Thread thread : threads) {
-                prepareThreadVOAndSaveIt(thread);
+                if (thread.getId() > 0)
+                    prepareThreadVOAndSaveIt(thread);
             }
 
         });
@@ -2583,7 +2726,6 @@ public class MessageDatabaseHelper {
     }
 
 
-    //TODO get from another thread
     //Cache contact
     @NonNull
     public Contact getContactById(long id) {
@@ -2604,7 +2746,6 @@ public class MessageDatabaseHelper {
         );
     }
 
-    //TODO could load from another thread
     @NonNull
     public List<Contact> getContactsByFirst(String firstName) {
         List<Contact> contacts = new ArrayList<>();
@@ -2656,7 +2797,6 @@ public class MessageDatabaseHelper {
         }
         return contacts;
     }
-    //TODO could load from another thread
 
     @NonNull
     public List<Contact> getContactsByFirstAndLast(String firstName, String lastName) {
@@ -2684,7 +2824,6 @@ public class MessageDatabaseHelper {
 
         return contacts;
     }
-    //TODO could load from another thread
 
     @NonNull
     public List<Contact> getContactByCell(String cellphoneNumber) {
@@ -2711,7 +2850,6 @@ public class MessageDatabaseHelper {
         }
         return contacts;
     }
-    //TODO could load from another thread
 
     @NonNull
     public List<Contact> getContactsByEmail(String email) {
@@ -2765,7 +2903,6 @@ public class MessageDatabaseHelper {
 
     }
 
-    //TODO could load from another thread
 
     public List<GapMessageVO> getAllGaps(long threadId) {
 
@@ -2787,8 +2924,8 @@ public class MessageDatabaseHelper {
         new PodThreadManager()
                 .doThisSafe(work);
 
-
     }
+
 
     public void getGap(long id, OnWorkDone listener) {
 
@@ -2952,7 +3089,6 @@ public class MessageDatabaseHelper {
     }
 
 
-    //todo could get from another thread
     private ChatProfileVO getChatProfile(long id) {
 
         return messageDao.getChatProfileVOById(id);
